Make phan happy

+++ projects/plugins/boost/app/modules/optimizations/lcp/class-lcp-optimize-img-tag.php
@@ -144,7 +144,7 @@ class LCP_Optimize_Img_Tag {
 
 				// If it's a Moto G Power, include a 1.75 DPR for accurate lighthouse representation of the optimized image.
 				if ( isset( $breakpoint['maxWidth'] ) && $breakpoint['maxWidth'] === 412 ) {
-					$dimensions[ round( $width * 1.75 ) ] = round( $height * 1.75 );
+					$dimensions[ (int) round( $width * 1.75 ) ] = round( $height * 1.75 );
 				}
 
 				// Include 2x DPR.
